# lib/api-zod/src/generated/types/createWallpaperBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LuminaWalls API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateWallpaperBody {
  title: string;
  /** @nullable */
  description?: string | null;
  imageUrl: string;
  /** @nullable */
  thumbnailUrl?: string | null;
  downloadUrl: string;
  category: string;
  resolution: string;
  width: number;
  height: number;
  tags?: string[];
  /** @nullable */
  fileSize?: number | null;
  isFeatured?: boolean;
  isAiGenerated?: boolean;
}