# lib/api-zod/src/generated/types/favoriteEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LuminaWalls API specification
 * OpenAPI spec version: 0.1.0
 */

export interface FavoriteEntry {
  id: number;
  userId: string;
  wallpaperId: number;
  createdAt: string;
}