# lib/api-zod/src/generated/types/generateWallpaperResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LuminaWalls API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Wallpaper } from "./wallpaper";

export interface GenerateWallpaperResponse {
  imageUrl: string;
  width: number;
  height: number;
  prompt: string;
  wallpaper: Wallpaper;
}